+++ Runtime/Scenes/Import.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using System.Threading.Tasks;
+using ComradeVanti.CSharpTools;
 using UnityEngine;
 using AssimpScene = Assimp.Scene;
 using static At.Ac.FhStp.Import3D.Meshes.Import;
@@ -9,9 +10,14 @@ namespace At.Ac.FhStp.Import3D.Scenes
 {
     internal static class Import
     {
-        internal static async Task<GameObject> ImportScene(AssimpScene scene, string filePath)
+        private static string MakeSceneNameFrom(string filePath) =>
+            Path.GetFileNameWithoutExtension(filePath);
+
+        internal static async Task<GameObject> ImportScene(AssimpScene scene,
+            string filePath, ImportConfig config)
         {
-            var sceneName = Path.GetFileNameWithoutExtension(filePath);
+            var sceneName = config.SceneNameOverride
+                .DefaultWith(() => MakeSceneNameFrom(filePath));
 
             Task<Mesh> ImportMeshWithIndex(int index) =>
                 ImportMesh(scene.Meshes[index]);